fix: show required token message, move between util files
