+++ src/stacks-hierarchy/Root_PurchaseOverviewScreen/Root_PurchaseOverviewScreen.tsx
@@ -298,11 +298,7 @@ export const Root_PurchaseOverviewScreen: React.FC<Props> = ({
             />
           ) : (
             <View style={styles.messages}>
-              {!isOnBehalfOfToggle && (
-                <PurchaseMessages
-                  requiresTokenOnMobile={requiresTokenOnMobile}
-                />
-              )}
+              <PurchaseMessages requiresTokenOnMobile={requiresTokenOnMobile} />
               <GlobalMessage
                 globalMessageContext={
                   GlobalMessageContextEnum.appPurchaseOverview
